xc-service-manage-course/src/test/java/com/xuecheng/manage_course/service/CourseServiceTest.java
package com.xuecheng.manage_course.service;

import com.github.pagehelper.Page;
import com.xuecheng.framework.domain.course.CourseBase;
import com.xuecheng.framework.model.response.QueryResponseResult;
import org.junit.Assert;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.junit4.SpringRunner;

import static org.junit.Assert.*;

@RunWith(SpringRunner.class)
@SpringBootTest
public class CourseServiceTest {

    @Autowired
    private CourseService courseService;

    @Test
    public void get_course_page_list() {
        QueryResponseResult courseBasePage = courseService.findPageCourseBase(1, 5);
        Assert.assertTrue(courseBasePage.getQueryResult().getTotal()>0);
    }
}